+++ library/src/main/kotlin/com/chuckerteam/chucker/internal/ui/transaction/TransactionPagerAdapter.kt
@@ -24,5 +24,5 @@ internal class TransactionPagerAdapter(context: Context, fm: FragmentManager) :
 
     override fun getCount(): Int = titles.size
 
-    override fun getPageTitle(position: Int): CharSequence? = titles[position]
+    override fun getPageTitle(position: Int): CharSequence = titles[position]
 }